refactor: move cast vote statistics queries
