+++ packages/windmill/src/postgres/cast_vote.rs
@@ -19,6 +19,106 @@ struct CastVoteAnnotations<'a> {
     voting_channel: VotingStatusChannel,
 }
 
+#[derive(Clone, Copy)]
+pub(crate) enum CastVoteRelation {
+    Production,
+    #[cfg(test)]
+    StatisticsTest,
+}
+
+impl CastVoteRelation {
+    /// PostgreSQL identifiers cannot be query parameters. Keeping the relation
+    /// selector closed prevents caller-controlled text from reaching the SQL.
+    fn sql_identifier(self) -> &'static str {
+        match self {
+            Self::Production => "sequent_backend.cast_vote",
+            #[cfg(test)]
+            Self::StatisticsTest => "pg_temp.cast_vote_stats_test",
+        }
+    }
+}
+
+pub(crate) fn count_distinct_voters_by_channel_query(
+    cast_vote_relation: CastVoteRelation,
+    filter_by_election: bool,
+) -> String {
+    let election_filter = if filter_by_election {
+        "AND election_id = $4"
+    } else {
+        ""
+    };
+    let cast_vote_relation = cast_vote_relation.sql_identifier();
+
+    format!(
+        r#"
+            WITH latest_valid_votes AS (
+                SELECT DISTINCT ON (voter_id_string)
+                    voter_id_string,
+                    COALESCE(annotations->>'voting_channel', 'ONLINE') AS channel
+                FROM {cast_vote_relation}
+                WHERE
+                    tenant_id = $1 AND
+                    election_event_id = $2 AND
+                    status = $3 AND
+                    voter_id_string IS NOT NULL
+                    {election_filter}
+                ORDER BY
+                    voter_id_string,
+                    created_at DESC NULLS LAST,
+                    id DESC
+            )
+            SELECT
+                channel,
+                COUNT(*) AS count
+            FROM latest_valid_votes
+            GROUP BY channel
+            ORDER BY channel;
+            "#
+    )
+}
+
+pub(crate) fn count_votes_per_day_query(cast_vote_relation: CastVoteRelation) -> String {
+    let cast_vote_relation = cast_vote_relation.sql_identifier();
+
+    format!(
+        r#"
+            WITH date_series AS (
+                SELECT
+                    (t.day)::date AS day
+                FROM
+                    generate_series(
+                        $3::date,
+                        $4::date,
+                        interval '1 day'
+                    ) AS t(day)
+            )
+            SELECT
+                ds.day,
+                COALESCE(v.annotations->>'voting_channel', 'ONLINE') AS channel,
+                COUNT(v.id) AS day_count
+            FROM
+                date_series ds
+            LEFT JOIN {cast_vote_relation} v ON ds.day = DATE(v.created_at AT TIME ZONE $5)
+                AND v.tenant_id = $1
+                AND v.election_event_id = $2
+                AND (v.election_id = $6 OR $6 IS NULL)
+                AND v.status = $7
+            WHERE
+                (
+                    DATE(v.created_at AT TIME ZONE $5) >= $3 AND
+                    DATE(v.created_at AT TIME ZONE $5) <= $4
+                )
+                OR v.created_at IS NULL
+            GROUP BY
+                ds.day,
+                COALESCE(v.annotations->>'voting_channel', 'ONLINE')
+            ORDER BY
+                ds.day,
+                channel;
+            "#
+    )
+}
+
 fn cast_vote_annotations(
     voter_ip: &Option<String>,
     voter_country: &Option<String>,
